requirements.txt
flask
gunicorn
requests
psycopg2-binary
python-dotenv